feat(filter)!: Change behavior for rules that fail filter criteria tests
Instead of "skipping" the rule will now fail. This aligns Rule behavior with how filters work through the rest of CM which should reduce cognitive load and development effort.

If the skipping behavior is still desired a user can use a RuleSet with OR condition to achieve the same effect.